lib: add basic capn config access protocol
This commit introduces the qzc.capnp schema file. This schema file is
only here for information, despite it is possible to automate c files
generation, from this scheme file.
In addition, qzc framework needs some server API to handle qzc queries
from any qzc client. Among the APIs, there is the possibility to handle
delete/destroy/unset operations. It is also possible to answer to get
operations that request some context to be retrieved. Finally, it is
possible to handle iterations over a table, for example.

Signed-off-by: David Lamparter <equinox@opensourcerouting.org>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>