bgpd: VRF import processing
This enhancement permits to export received or locally set route
entries for MPLS VPN, to the appropriate VRF. By introducing the
RFC4360 and BGP extended communities, some export route target are
received on BGP update, or some import route target are set on a
local route entry. It results from exporting the current route
information to the appropriate VRF RIB entry.
This commit is relying on BGP VRF processing, and supports VRF
RIB election as for standard RIBs. This makes it reusable, even
for multipath feature.

Signed-off-by: Christian Franke <chris@opensourcerouting.org>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>