bgpd: default originate for VPNv4 and ENCAP address family
This commit introduces default route facility for VPNv4 and ENCAP
subsequent address family.
When user sets a default route with a VPNv4 or ENCAP address family
to any neighbor, Route Distinguisher must be taken into account to fill
the MP_REACH_NLRI attribute of UPDATE message accordingly.
Reversely, when user withdraw a default route with a VPNv4 address family
to any neighbor, Route Distinguisher must be taken into account to fill
the MP_UNREACH_NLRI attribute of UPDATE message accordingly.
Since the default route is no stored in RIB, it must be stored in struct
peer to be able to send default route to peers that restart with support
of VPNv4 address family. When sending routes to a restarting BGP speaker,
there are as many UPDATE messages sent as the number of default route
for all RD configured for that peer.

Signed-off-by: Julien Courtat <julien.courtat@6wind.com>