qthrift: thrift support for createPeer, deletePeer feature
BGP createPeer() and deletePeer() thrift API support. The
implementation relies on zmq connection with bgpd. a createchild
and set operation is performed to create new peer context. Pre-check
that neighbor is not already created, is done against a peer caching
mecanism in qthriftd.
Once created, neighbor is configured for (afi,safi)=(IP, MPLS_VPN).
This operation relies on setting a table in peer context. For that,
as capnproto messages do not copy the whole table, a context matching
the wished (afi,safi) context is passed in the message. It permits
bgpd to enable directly the correct flag in the bgp peer (afi,safi)
table.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>