qthrift: increase asNumber from 32 bits to 64 bits
In Thrift, parameters are always signed, thus limiting the maximum
AS number from 1 to 2147483647 (2³¹-1). Using an AS number encoded
on 64 bits will permit to use all the 32 bit positive values
from 1 to 4294967295.
This commit introduced new thrift API changes, as well as files
generated by thrift compiler, as well as the changes in bgp
configurator API.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>