bgpd: VPNv4 default route support in Graceful Restart
Since the default route is no stored in RIB, it must be stored in struct
peer to be able to send default route to peers that restart with support
of VPNv4 address family.
When sending routes to a restarting BGP speaker, there are as many
UPDATE messages sent as the number of default route for all RD configured
for that peer.

Signed-off-by: Julien Courtat <julien.courtat@6wind.com>