bgpd: import/export algorithm
On receiving BGP updates from remote BGP speakers, NLRI is analysed, and
the message to add/remove to the global RIB, is also sent for processing
to the BGP VRF RIB tables. The same procedure is done for statically
(un)configured or updated route entries.
In order to send the message to the associate VRF RIB tables, the list
of RD that subscribed to the receiving RD, as well as the exported RD of
the received BGP update are parsed.
The BGP information structure to be changed is compared against the BGP
VRF RIB route node entries. Consequently, the bgp best selection
algorithm is applied, including multipath route selection algorithm.
Consecutive to the new election, a message of update or withdrawal is
sent, to signify a change in each VRF RIB.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>