bgpd: if multipath entries reached, no changes in routing
In order to propagate information for any multipath changes, the flag
MULTIPATH_CHG is used to notify that there are changes or not. Like
that, not only aggregation update will not be needed, but also, as
no multipath changes or any attribute changes are seen, then the bgp
processing does not perform anything.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>