bgpd: update notification message for evpn entries updates in VRF RIB.
When importing EVPN entries to VRF RIBS, special processing is done.
Labels are processed differently according to the VRF layer type.
In addition to this, routermac will not be displayed if VRF layer type
is layer2 In addition to this, a log event is displayed to give
information on MAC/IP and RT5 information.

Signed-off-by: Julien Courtat <julien.courtat@6wind.com>