qthrift: changes monitoring mecanism to check for bgp updater connection
Not destroying or closing thrift bgp updater connection implies that
tcp connection is still present on the client point of view. The
monitoring mecanism was ineffective, while the reset of the connection
was not triggered due to wrong status. The monitoring mecanism resets
thrift bgp updater connection each 5 seconds to check that. Since all
calls are syncronous, and qthrift uses only one thread to communicate
with thrift server, reset should not cause a problem.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>