lib: enhance str2esi and str2mac functions
Those functions are used to check that input format of an ESI or a
mac address passed pas paramter is correct or not. Accepted format
are for mac address aa:bb:cc:dd:ee:ff. Accepted format for esi is:
00:11:22:33:44:55:66:77:88:99. Accepted format for mac address is:
aa:bb:cc:dd:ee:ff.
In addition, three other functions are used to reverse back, that
is to say to give a displayable string from an input standing from
the internal view of ESI or mac address. Note that 2 translation
functions are available for mac address, one taking as input
internal mac address format, the other taking as input extended
router mac community.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>