bgpd: add capnproto support for prefix definition
Prefix conforming to RFC 7432 (ethtag, mac_len/mac, ip_len/ip) is build
based on mac address and/or IP address and/or ethernet tag.
Because prefix is a structure used in many parts of the capnproto
interface, the commit handles those modifications.
onUpdateWithdrawRoute() are modified accordingly.

The prefix length value returned by BGPd stands for the whole prefix
length value, including mac address and ethernet tag, when address
family is AF_L2VPN. In this specific case, only host IPs are expected
and the prefix length to be returned to capnprot client is the IP
prefix length.

Signed-off-by: Julien Courtat <julien.courtat@6wind.com>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>