bgpd: add VRF field layer type
On VRF creation, an additionnal parameter layer type can be provided to
distinguish between layer 2 and layer 3 vrf. layer 2 is used by EVPN
when dealing with MAC/IP route advertisement. layer 3 is used by any
other route L3VPN or EVPN route advertisement. An utility function is
proposed in order to extract the label associated to the layer of the
VRF. Actually when MAC/IP advertisements are exchanged for route
processing, when importing the processing in VRF, according to the VRF
layer, only the relevant label information will be collected.

Signed-off-by: Julien Courtat <julien.courtat@6wind.com>