bgpd: reception of auto discovery and vrf import processing
The auto discovery algorithm is depicted more in detail, when impacting
the VRF RIB table. For each RT set from the A/D, the prefix matching the
correct ESI and ethtag is parsed. If a match is found, and the A/D is a
withdraw message, then the entry is removed from ADJ-RIB-IN, as well as
VRF RIB table. If a match is not found, and the A/D is an add message,
then a new entry is appended and sent to bgp_process() for further
processing.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>