qthrift: (push/withdraw)Route of EVPN RT1 route advertisement
Detect L2VPN Auto Discovery messages from pushRoute and withdrawRoute
thrift messages. The detection is based on the fact that the protocol
type is EVPN and prefix is null.
According to that filter, sanity checks are done before sending the
corresponding message to bgpd through capn'n proto.
Note that this commit requires the qzc to be updated in order it returns
error code whitin it's set_element function.

Signed-off-by: Julien Courtat <julien.courtat@6wind.com>