bgpd: factorize calls to action for an entry in VRF RIB
When doing vrf import processing, on VRF RIBS, there may be an entry
to add or remove. When adding, a debugging informational trace is
triggered, and bgp_process is called for route election. When removing,
the entry is removed, then debugging informational trace is triggered.
In addition to this, adj-rib-in table per VRF is updated.

Additionaly, when receiving a/d message from a peer, the duplication of rib entries
may be huge. However, multipath is still applied by bgp process mecanism
after each new entry. In order to avoid useless calculation, bgp process
is done after having parsed the route node entry.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>