lib: check return code in qzc setelement
Even if functions invoked by capn'proto via the 'set element' interface
fails, the rpc caller is not noticed of the failure.
This patch is checking the return code on the remote program and
transfers it to the caller in the same way of get element.
Note that only the status of the function being invoked is get back to
the caller, not the capn'proto reply message, as the get element
interface does.

Signed-off-by: Julien Courtat <julien.courtat@6wind.com>