bgpd: for each rx a/d, ctxt is created for exported vrfs
To ease search in auto discovery contexts, received context is
references in each vrf mentioned in rt set. This includes new entries to
populate. It just duplicates context. It does not handle the removal of
a ad context, due to a lose of BGP session.

Because there must be a distinction between route distribution and vrf
import processing, a new list has been done to handle vrf processing per
vrf. In addition to this, ad_lookup routine is reused when RD is not
given as paramter.

This fix also permits to delete entries in VRF RIB associated, then delete
A/D internal structure. If not done, entries associated to A/D
MP_UNREACH message are not deleted.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>