bgpd: prevent from unlocking route node when attempting to process A/D
When updating VRF RIB against reception of an A/D message, the route
node lock shall be used carefully, in order to avoid crash. This is the
case when A/D messages are consecutively received and create/delete
entries. To avoid this, the processing is done within the route node
lock up of the calling function.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>