bgpd: inform if new entry is coming from local config or from remote BGP
In order to give information when an entry is added/removed from VRF
RIB, then the tracing information is improved. It permits to know if
the route configured is local or not. The word mngr is used to mention
either the configurator ( for configuring routes via vty), but also the
manager that is behind the capnproto interface ( which is called here
capnproto manager, because it does not only does configuration, but also
receives notifications).

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>