*: force the linker to not check for unresolved symbols at link
This commit bypasses the link check in case a shared library is to be
used in conjunction with one of the quagga daemons. This is the case of
zero ma library that may or not be available on some systems. Adding
-Wl,-unresolved-symbols=ignore-in-shared-libs to the linker flags
permits compiling without having to install zmq.
It also adapts vtysh scripting, as the capnproto and zeromq may be needed,
the compilation flags is added.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>