bgpd: bgp nexthop attribute is filtered against local IP addresses
If a BGP update is received with nexthop attribute standing for one of
the IPs of the local BGP speaker, then the BGP update message should be
rejected, which was not the case.
This handling is implementing extract from rfc4271:
A route originated by a BGP speaker SHALL NOT be advertised to a peer
using an address of that peer as NEXT_HOP. A BGP speaker SHALL NOT
install a route with itself as the next hop.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>