bgpd: enhance bgp evpn auto discovery utilities
This commit enhances an internal structure used to store the
configuration orders of evpn ad emitted by the router, and the
a/d messages received from remote. There is also a set of tools to
populate the structure and to display some internal contexts. Finally,
some other tools are available to forge an attribute,derived from auto
discovery settings.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>