bgpd: handle auto discovery import processing
This commit brings the feature in order to spread mac availability or
disappear on VRF RIBs. This feature is handling received auto discovery
messages, operates the import processing to get the list of VRFs ( by
means of route target), and on each VRF, a duplication or suppression of
entries in VRF RIB is done according to ESI or {ESI,Eth Tag}.
Note that RFC7432 stipulates that the entries of global RIB should not
be removed, before all MAC/IP corresponding are received just after auto
discovery message. This operation is just speeding up the procedure
related to VRF RIBs, and is conform with RFC7432 since it impacts only
VRF RIB.
Note also that the Auto Discovery import processing lasts all the
lifetime of the BGP session, meaning that if an A/D is received at T1,
if a MAC/IP to duplicate is received consecutively, then the duplication
related to this A/D will be carried out.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>