bgpd: add labeled unicast safi parameter
This commit add the support of labeled unicast safi parameter.
Some assumptions are taken into account. For instance, as there is no
route distinguisher attribute per entry, the rib tables are parsed like
IP unicast tables. On the contrario, Labeled unicast is considered as an
MP-BGP extension, and as such, it is able to support IPv4 or IPv6
nexthop attribute. For that, mp_nexthop attributes will be used to be
displayed or extracted for emission.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>