bgpd: check afi/safi with multipath is extended to vrf
The function that checks if multipath is enabled or not is modified to
check, upon a passed route_node argument, if that route node belongs to
a VRF and if that VRF has multipath configured.
This permits for VRF RIB processing to apply the same bgp best path
algorithm as the one that could apply for standard afi/safi.
When maximum path is set, in some situation, when attempting to install
too many multipath routes, the vrf where the install is done, has not
been found. Consequently, the entry was considered as multipath, whereas
it should not have been. This commit finds the correct vrf.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>