bgpd: change bgp_static structure for gateway_ip from ipv4 to ipv4/ipv6
The bgp_static structure is changed so that it is possible to support
both ipv4/ipv6 gateway IP. Note that the nature of the gateway IP
parameter depends of the nature of the prefix. If prefix is IPv6, then
the gatewayIP is IPv6. The same for IPv4. This is not possible for other
combinations.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>