bgpd: on no vrf matching, incoming vpnvx/evpnvx messages are hidden
If import processing fails due to no matching rt set found, and then
inability to add entries into a vrf rib, then the global RIB will have
the entry hidden, the entry will not appear.
If a new vrf is to be introduced, then the list of hidden entries from
vpnvx will be parsed and checked against new vrf. Then the entry will
appear in case.
Reversely, if a vrf is to be deleted, the associated entries in global
RIB will be hidden if no other vrf is being targeted by this entry.
If a static entry is under the way of being exported, then the entry in
the global RIB should be kept not hidden, independently if there are rts
or not, if it is a static entry locally configured.
Fix the show function handling of hidden entries. If an hidden entry was
detected, the function was exiting, wheras the expectation would have
been to continue on next iteration.
In order to clarify, a new vty command is available: show ip bgp vpnv4
all hidden. It will display all entries of global RIB.
In the case an hidden entry with best selection algorithm is picked up,
and is to be deleted, then the change is removing the entry.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>