bgpd: fix complement for rtset filtering feature
Fix the show function handling of hidden entries. If an hidden entry was
detected, the function was exiting, wheras the expectation would have
been to continue on next iteration.
In order to clarify, a new vty command is available: show ip bgp vpnv4
all hidden. It will display all entries of global RIB.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>