bgpd: update vrf entry when message received quicker than processing
When one receives one withdraw and one update message very quickly, the
VRF table entry is not updated accordingly.

This commit is to undo the effects of a previous call to bgp_info_delete
for VRF entry. The flags of BGP_INFO_REMOVED and BGP_INFO_VALID are
restored to bgp_info.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Liu Xiaofeng <xiaofeng.liu@6wind.com>