bgpd: support L3VPN dual stack
VRF RIB table can be enabled/disabled for IPv4 or IPv6 separately.

The VRF processing will be ingored for incoming update messages if the
corresponding RIB table has not been enabled. And in global RIB the
route entry will be hidden.

When one VRF RIB table is enabled, the hidden entries in global RIB
will be checked to see if they should be activated.

When one VRF RIB table is disabled, the entries against the same afi/safi
in global RIB will be checked to see if they should be set to be hidden.

Signed-off-by: Liu Xiaofeng <xiaofeng.liu@6wind.com>