qthriftd: add two thrift APIs enableEORDelay and sendEOR
Two new thrift commands are added:
sendEOR() and enableEORDelay()

In addition to this, in startBgp handler, original bgp parameters
such as bgp->v_update_delay should be retrieved so as it will
not be overwritten.
Also, as capnproto return code is not handled for EOR, ignore
that error code. As consequence, the qthrift code then should ignore it.

Signed-off-by: Liu Xiaofeng <xiaofeng.liu@6wind.com>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>