bgpd: enhance rfc7313 support
This commit introduces the RFC7313 capability.
This includes handling of BOR and EOR bits contained in the BGP route
refresh packet.
The rfc7313 capability must be propagated into all bgp code.
That capability will be linked to the EOR and BOR bits ( contrary to
what it is currently).
Then, some vty must be done to disable rfc7313 on request.
By default, rfc7313 capability will be sent, and will replace
NEW_REFRESH ( rfc2918) if peer supports rfc7313. Otherwise, fallback to
rfc2918 behaviour.
As this work is not totally finalised, but is used to enhance first BGP
refresh mechanism, the feature is kept but disabled.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>