bgpd: separate stale flag from stale refresh flag
this separation of flag in flags of each entry is here to avoid the case
where a graceful restart occur in the meantime that a BGP refresh
exchange occurs. This introduces a new internal flag:
BGP_INFO_STALE_REFRESH, that is updated only in the BGP refresh
mechanism used in rfc2918 and rfc7313.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>