qthriftd: initialize qthrift context with bgp as, when startBgp ok
An optimisation is done, in order to avoid the capnproto calls to be
done too frequently with sartBgp() and delayEor(). The startBgp call
will set the qthrift context to init ( with as Number) only at the end
of the procedure.
In addition to this, if delayEor() is called, the configuration is
stored in qthrift and injected during the startBgp() call.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>