bgpd: Fix AS_PATH size calculation for long paths
If you have an AS_PATH with more entries than
what can be written into a single AS_SEGMENT_MAX
it needs to be broken up.  The code that noticed
that the AS_PATH needs to be broken up was not
correctly calculating the size of the resulting
message.  This patch addresses this issue.

(cherry picked from Quagga Routing Software commit 7a42b78b)
Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>