bfdd: optimize debounce timer
The debounce timer is used to bounce off network flaps. It should be
used only when the BFD state is changing too quickly. If the state
changes once from UP to DOWN or from DOWN to UP, the debounce timer
should not be triggered.

This commit optimizes the debounce timer to implement the above
objective.

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>