bgpd: process Control Plane Independent flag
When graceful restart is disabled for a bgp peer, Control Plane
Independent(C-Bit) flag is set while adding a bfd neighbour.
So that bfdd can set the C bit in bfd hello message.

On the other hand, different process of BFD_NEIGH_DOWN message
is performed if peer is gracefully restarted:

- C-Bit is cleared
  In this case, BFD_NEIGH_DOWN message is ignored. And the routes
  learned from peer are held until the graceful timer expires.

- C-Bit is set
  In this case, graceful restart is aborted. And the routes learned
  from peer is flushed immediately.

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>