bgpd: add bfd neighbour even if bgp connection is not up
When a bfd neighbour is added, local address and remote address
, together with the local ifindex, will be encapsulated in
BFD_CNEIGH_ADD message. Also, local address and the interface on
which BFD packets are transmitted can be confirmed when bgp
connection is UP. That's why a bfd neighbour is added by bgpd
only when the corresponding bgp connection is established.

But there is an issue in the following case:

A. In the initiation, BGP connections and bfd sessions are built
   between local side and remote side.
B. For some reason(e.g. power is off), remote BGP peer cannot be
   reached.
C. Local bfd regards that peer is down after the detection time.
   And on local side BFD neighbour state is Down and bgp neighbour
   state is changed to "Idle" which means it does not send BGP OPEN
   messages and does not respond to BGP OPEN messages.
D. After some time, remote peer is restored. But bgp connection
   cannot be established because local bgp neighbour is in "Idle"
   state.
E. And on remote, bfd neighbour cannot be added for bgp connection
   is not established. So on local side, bgp neighbour is always
   in "Idle" state.

To fix the issue, bfd neighbour will always be added even if bgp
connection is not established. The local bfd address and ifindex
are set to 0 in BFD_CNEIGH_ADD message when BGP connection is not
UP. This makes it possible to have bfd neighbour in bfdd before
BGP connection is UP. Once bfd detects peer liveness, BGP neighbour
is changed to "Connect" state and starts connecting peer.

If bgp connection is established later, local address of bfd neighbour
in bfdd will be updated to the source address used by bgp connection.

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>