bfdd: make it possible to add a bfd neighbour without local address
This commit makes bfdd have the ability to add a new neighbour without
local address and ifindex.

When bgp neighbour is created in bgpd, bfd neighbour is also added.
But local address is an unspecified address(0.0.0.0 or ::/128), and
ifindex is 0. Bfd session is started to detect peer status.

Once bgp connection is established, the corresponding bfd neighbour
is also updated. Local address is changed to the source address
used by bgp connection.

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>