bgpd: only set C-bit for established neighbor when adding a bfd neighbor
When bgp peer is added, a bfd neighbor will be added to bfdd in
order to start bfd session. Before this commit, C-bit will be
set in the case graceful restart is not enabled, whatever bgp
connection is established or not. Now C-bit is set only when
bgp connection is established.

Also, bfd neighbor will be added with unspecified local address
if bgp connection is not established.

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>