bgpd: set min refresh timer of 5 seconds
This is a workaround that happens to swap the selected entries in BGP
VRF RIB entries. As STALE status is set only on global VRF RIB, the VRF
RIB entries have no information about the selection or not of the entry.
As consequence, there is a gap between both RIBs, and bgp best selection
algorithm acts differently. To lower the issue, the refresh timer
triggered is set to a minimum of 5 seconds, so that bgp neigbors may
have time to reply. Indeed, refresh timer is syncronised with graceful
restart, as the resulting STALE status should have the same effect. But
for refresh, more time has to wait. A more elaborate solution will have
to be found, in order to get the STALE status on VRF RIB too.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>