bgpd: add support for BGP bestpath selection deferral
This commit adds a selection deferral timer to support
the case BGP peer never sends EoR marker. When a peer is
connected, the selection deferral timer is launched.
At the expiration of the timer, BGP bestpath selection
will be done even EoR marker is not received. The default
selection deferral timeout to 360 seconds.

If deferral timer is set to 0, then process rt entries
as soon as possible. That behaviour was the bgp behavior
until now, despite this behavior was not rfc compliant.
This behavior is kept in case some side effects appear.

New vty command is added to configure the timeout value in
miliseconds:

  bgpd(config)# bgp bestpath selection-deferral <0-720000>

New bgp instances will inherit from that deferral value.

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>
Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>