bgpd: api route get returns selected entry or mpath entry
When facing multiple path entries, there were cases where the first
selected entry entry was not belonging neither to the selected multipath
entries ( eg: if 3 ecmp entries and only 2 selected), nor to the
selected entry ( eg: if 3 ecmp entries and only 1 selected). The fix
consists in returning either a selected entry or a mpath entry.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>
Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>