bfdd: support bfd underlay limit
CSC connected to 2-DC-Gwy's in the OverLay, but in UnderLay
both the DC-Gwy's are going through the same switch. If the
UnderLay-switch goes for a reboot and it takes time(~20seconds)
to come back-up. This is leading to BFD timers expiry, in
turn we get peer-Down for both the neighbors:
 - all peers get down
 - all routes are withdrawn
 - the data path traffic is interrupted.

To avoid such scenarios, in this commit, peerDown() event is
sent only when:

  floor(number_of_peers_AVAILABLE) >= ceil(number_of_peers/2

Following command enables/disables the service using vty command:

  bfd(config)# [no] bfd underlay limit

The peer down event is not blocked permanently. First, because
BGP may detect that remote peer is not available anymore. By
default, the peer down event is blocked during 180 seconds, that
is to say the time that an underlay switch needs to restart. Also
that value is equivalent to the default bgp hold timer. To change
that switch timer, use the following command :

  bfd(config)# [no] bfd underlay limit timeout [<0-65535>]

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>