bgpd: relaunch bgp, when adding bfd
bfd add action feedback is sent back to bgp zebra layer. by default, the
status of bfd becomes up; but it is necessary to relaunch BGP session in
case the BGP session was down. In case of BFD forced BGP to be down, the
previous status was Idle. Then a reconnect will force BGP to rety to
connect.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>