bgpd: fix vrf process when update message arrives before withdraw process
When there are a lot of routes from peer and "clear ip bgp *" is
performed, maybe an update message comes quickly after withdraw
message from the same peer. The withdraw process is finished for
global rib, but not finished for vrf rib. In this case,
onUpdatePushRoute message is missed to be sent to ODL, because
the perfix has been scheduled for withdraw process. This commit
is to undo the effects of a previous call to bgp_info_delete for
VRF entry. The flag BGP_INFO_REMOVED is unset and BGP_INFO_VALID is
set.

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>