bgpd: postpone bgp connection to one peer if a same peer is dying
When a peer is deleted and there have been a lot of routes learnt
from this peer, it will take much time to withdraw the routes
before this peer is freed. During the time, if the same peer is
created, old peer probably can not be freed, in the meanwhile,
some onUpdatePushRoutes message may be missed to be sent when
update messages are received after new peer is established.

The reason is that both withdrawing and updating processes are
performed on route entries in global rib and vrf rib. Withdraw
process is for old peer, and updating process is for new peer,
which are mixed and disorderly.

This commit postpones the bgp connection to new peer until old
peer is freed.

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>