bgpd: permit to reprocess entries with previously failed exportations
if a vrf has not yet its afi/safi provisioned, and an exportation is
being scheduled for this vrf, then the exportation will fail. when the
provisioning for the afi/safi happens, it will be possible for the
entries to be reparsed, and if there is a change in the number of
targeted vrf that fail, then the bgp processing will take place on that
vrf.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>