qthriftd: ability to configure stalemarker timer
in case of breaking of the connection from qbgp to odl, this may signify
that an odl move within a cluster may take place; a timer is run, in
order to remember to remove the non unmarked entries ( vrfs, and
prefixes). That timer will generally stick with end of rib, that is to
say either the one defined by sendEorDelay() or when sendEor() is
called. In order to give more flexibility, the timer by default is set
to 1800 Seconds maximum ( which is supposed to be the same as Eor
delay), but can also be changed, by using option '-M <1-3600>' value.

Signed-off-by: Philippe Guibert <philippe.guibert@6wind.com>