bfdd: schedule delete timer with 1000 milliseconds delay
Once a neighbor is deleted, the neighbor is not deleted immediately,
while a delete timer is scheduled. Before this commit, the delay
is MSEC(neighp->negtxint * neighp->lmulti), which is too long when
neighp->negtxint is set to a big value.

This commit sets the delay to 1000 milliseconds whatever neighp->negtxint
and neighp->lmulti are.

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>