bfdd: don't start detection timer if admindown timeout timer is activated
When admindown timeout timer is activated, detection timer should not
be triggered.

Both admindown timeout timer and session timeout timer have the ability
to reset BFD timers so that BFD negotiation can be started again.

The admindown timeout timer is expected to handle the case when
AdminDown message is received from remote peer. And session timeout
timer is expected to handle the case when remote peer is detected down
by detection timer.

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>