bgpd: support EVPN MAC mobility
This commit implements EVPN MAC mobility feature.

The implementation assumes that MAC/IP binding remains the same after
MAC Mobility. Other scenarios where a MAC move results in a different
MAC/IP binding are not supported.

For EVPN RT2 advertisement, if MAC sequence number is not zero, a MAC
Mobility Extended Community with the MAC sequence will be advertised.

When a local RT2 entry is configured, if there is already a remote
entry received with the same MAC/IP, a new advertisement, with a MAC
sequence which is one greater than the MAC sequence of remote received
entry, will be advertised;

When a RT2 advertisement with a high MAC sequence is received from
remote, if there is already a local configured RT2 entry with lower
MAC sequence, the local configured entry should be withdrawn;

For best path selection algorithm, the entry having higher MAC sequence
is preferred; if two entries have a same MAC sequence, the entry
advertised by the BGP peer with lower IP will be preferred.

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>