lib/vty: add a function to reset all VTY status except current vty
This commit introduces a new function vty_reset_other_vtys(). Compared to
vty_reset(), the new function is to reset all VTY status except current
vty.

The function could be used to be called by a vty command which has to reset
all VTYs. Current vty will be accessed by vty_execute() after the command
execution, so it should not be closed during the command execution, while
it could be closed in vty_read().

Signed-off-by: Xiaofeng Liu <xiaofeng.liu@6wind.com>